Modify Util and remove GenerateAppProperties
